Make examples in readme more minimal
In research we’ve found that what developers need is some simple
examples that they can copy and paste to get started with.

Giving each example a heading helps as well, because it breaks the
process down into a series of steps.

This follows the approach we’ve taken for the Python client:
https://github.com/alphagov/notifications-python-client/blob/2c3c51d82b2d17f50d9a1d4dc327dca277454952/README.md