Fix broken client initialisation
The wrapper as always passing three arguments to the constructor. eg
`new NotifyClient('abc123')` would result in `ApiClient` getting called
with (`abc123`, undefined, undefined).

Because we count the arguments in `ApiClient` to determine what they
mean this was completely broke the client, because it was trying to
use the API key as the base url, `undefined` as the API key and
`undefined` as the service ID.

This commit does some Javascript gymnastics to always call the
constructor with the same number of arguments that the wrapper is given.

Technique taken from: http://stackoverflow.com/a/8843181/147318

> A bit of explanation: We need to run new on a function that takes a
> limited number of arguments. The bind method allows us to do it like
> so:
>
> ```
> var f = Cls.bind(anything, arg1, arg2, ...);
> result = new f();
> ```

The `slice` call is necessary because `arguments` isn’t a proper array.