Change sendSms function to take an options object
sendSms previously took five parameters, three of which were optional.

The function has now been changed to take an optional `options` object
which can contain `reference`, `personalisation` and `smsSenderId`.
This means that there is now no need to pass in `undefined` if only some
of the optional parameters are needed.