Fix randomly failing unit tests by mocking date
The date was not being mocked in the api_client.js and notification.js
unit tests, which was causing these to randomly fail. This is because
the date gets used to generate a Notify token for the request header. By
using the real date, the request that we mocked using Nock and the request
that we attempted to make would occasionally have different headers.