Add typings to the NotifyClient
I'm introducing types for TypeScirpt to mainly satisfy the linters and
compilers, but also to provide extra code documentation that can be
previewed on the fly.

PaaS is using Notify and TypeScript for some their applications. Thus
far, we got around with typings in our local repo but I think other
teams that might be using TypeScript could benefit from this.

There is one extra file, which isn't actually a `spec`, but rather proof
that the types work as expected.