Bump node version we run our tests against.
Node v6 is end of life. Node v8 is still supported, although will
become end of life at the end of the year.

On the assumption that our users should not be running the client
on versions of node that are EOL, we should run our tests on
a node version that is still supported. I've bumped to v10 which
still has a fair amount of lifetime left.

Arguably, we should test against multiple versions of node but
this commit is not trying to do too much.