Fix node client integration tests on Jenkins
Recently, we have changed the Dockerfile in this repo, so it conforms
with our new Concourse pipeline. Yet, this made it incompatible with
Jenkins, and it turned out we are still running integration tests
on Jenkins when deploying the API.
So now we will have two Dockerfiles to help both Concourse and
Jenkins jobs run smoothly in the transition period.