BAU: remove __dirname global call from api client
`__dirname` is a global value made available to Node at runtime. This
will provide the directory where the exucuting file is located. As the
client files are static and always run from the same context, relying on
`__dirname` should not be required here.

Removing this path concatination would also resolve a corner case if the
library is being built by a code bundler that replaces the global value
of `__dirname` (this is an edge case and not default behaviour however may be
set by some configurations).

If the files are being bundled, directly referencing the relative
package JSON independent of the `__dirname` value will allow the bundler
to include this file.

More information on why this global may be overriden by a bundler can be
found here https://webpack.js.org/configuration/node/#node__dirname.