Fix Snyk prepublish command
* `prepublish` is deprecated and has been replaced with `prepare`,
so this changes the command in `package.json`
* npm commands, such `prepare` which is now used by Snyk, can't be run
as root. This adds a new user to the Dockerfile that can be used to run
them. We can't switch to this new user in the Dockerfile because we
still need to run as root when testing on Concourse, so we just use the
new user when running the command `package.json`