Add missing comma to variable declaration
Spotted in
https://github.com/alphagov/notifications-node-client/issues/137
(thanks @neilbmclaughlin), this missing comma
meant:
1. the code on line 5 was made into its own
variable declaration*
2. that declaration was made accidently global by
the lack of a `var` prefix

Having `version` as a global doesn't seem to break
anything but I don't think it's intentional so we
should change it.

This adds it back, making lines 1-5 a single
statement again.

*I'm pretty sure this is due to JavaScript's
Automatic Semicolon Insertion (ASI) adding a
semicolon at the end of line 4, splitting line 5
off from the `var` statement that started on line
1. I think that happened because the `var`
statement doesn't allow the End Of Line (EOL)
character at the end of line 4 without it being
preceded by a comma. That would trigger the first
rule of semicolon insertion, described here:

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Lexical_grammar#Automatic_semicolon_insertion