Remove redundant build-with-docker task
This isn't referenced anywhere or used by CI, which builds its own
Docker image. We do need to run it to install packages, but there's
no reason why it needs to be separate from "prepare-", which we'll
rename in a later commit. This combines the tasks together.

Note that the previous task tried to run "build" locally, which may
not have worked if NPM wasn't installed on the host machine. It's
unclear why a Docker-specific command tried to do this, since we
shouldn't be doing anything on the local machine.

Note also that "-C docker" did nothing (it just changed the working
directory of the command, which had no effect).