Add package-lock.json file of current state
The package-lock.json file will not be used by consumers of the
node client as per 'The difference is that package-lock.json cannot
be published, and it will be ignored if found in any place other
than the root project.
https://docs.npmjs.com/cli/v8/configuring-npm/package-lock-json

It will though be used by developers developing locally and by our
CI, meaning that we can guarantee which version of software is
being used. This is particularly important for devs and CI as a
malicious package, installed unintentionally, could compromise
the machine and access additional things it shouldn't. This was
something that was highlighted after some high profile package
takeovers:
https://portswigger.net/daily-swig/popular-npm-package-ua-parser-js-poisoned-with-cryptomining-password-stealing-malware