Add support for new send a file features
Adds client support for the new 'confirm email before download' and
'custom retention period' features when sending a file by email.

Bumps the client to v5.2.0

Update the `prepareUpload` function to take an `options` parameter. This
may just be `true/false` in some cases, and if that's the case it
represents the `isCsv` parameter. But it also may be an object
containing `isCsv` or the new doc download features,
`confirmEmailBeforeDownload` and `retentionPeriod`.