Support strings in prepareUpload
We currently only support Buffers as the value for `file` passed into
prepareUpload (this is the default value returned from fs.readFile with
no encoding specified). However, if an encoding _is_ specified (eg
binary), then `fs.readFile` instead returns a string.

If a string was passed into `prepareUpload` before, then
`toString('base64')` would fail and just leave the object unchanged. If
we receive a string, let's wrap it in a buffer and then convert it to
base64 as expected.