Fix global variables
Aa type error ('content is not defined') was thrown when using this
library as a dependency in a Typescript project, in strict mode, in
combination with esbuild bundling. Esbuild respects the 'alwaysStrict'
setting in a local tsconfig, and so a 'use strict;' directive is
inserted which affects all code in the bundle. Assigning to an
undeclared variable is an error when the directive is in place.