Bump minimum axios version to 1.2.0
The axios 1.x branch has several security vulnerability fixes.

Because we are on the 0.x branch consumers of our client can’t upgrade
to fix the vulnerabilities.

The current version of Axios is 1.6.1. This commit sets the minimum
version to 1.2.0. This is the lowest version of the 1.x branch where our
tests still pass.

Setting it to the lowest possible version means that:
- consumers who want the latest version can install it
- consumers who don’t want to or can’t upgrade from an older version can
  still use our client (as long as they have at least version 1.2.0 of
  Axios)

For some reason `protocol` is now a required property of `proxyConfig`
at least in our test scenario. So this commit also updates the tests so
they still work.