Allow setting custom filenames for documents
Update client for the new document download feature which allows
specifying a download filename for files sent by email.

This supercedes the old `isCsv` parameter that was inconsistent at best.

Given this, we are dropping support for `isCsv` altogether which lets us
clean up the interface for the `prepareUpload` function to now only use
the new options dict.