Update integration tests to work in FT environment
Also removes some assertions on the content of the templates. We already
assert that the returned JSON has the correct schema, so these
assertions served only to increase coupling between the tests and the
fixtures, without improving coverage.

Also bumps the number of retries when encountering a PDFNotReadyError,
because the dev envs seem to take longer to process them than preview
does.